requirements.txt
aiohttp==3.7.3
alembic==1.5.2
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
blinker==1.4
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
discord.py==1.6.0
dominate==2.6.0
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.2.0
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
redis==3.5.3
requests==2.25.1
rq==1.7.0
six==1.15.0
SQLAlchemy==1.3.22
typing-extensions==3.7.4.3
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.3
yarl==1.6.3
